Fixed capturing user voice
Before users could only speak for 10 seconds. This was the most optimal way and also would keep listening after user stopped talking for like 3 seconds. Making the program seem like it had a significant delay. It no longer works like that. It now listens to the user as long as they are speaking. The program will wait for two seconds to consider the user to be done speaking, then translate. The pprogram will also wait for one second without hearing anything, to consider the speaker to be pausing. This creates for a 3 second period where the user can catch their breathe or end their statement.